# Implement `top` option in queryFTS