--- a/extension/fts/src/function/fts_config.cpp
+++ b/extension/fts/src/function/fts_config.cpp
@@ -185,8 +185,8 @@ QueryFTSConfig::QueryFTSConfig(const function::optional_params_t& optionalParams
         } else if (Conjunctive::NAME == lowerCaseName) {
             value.validateType(Conjunctive::TYPE);
             isConjunctive = value.getValue<bool>();
-        } else if (Top::NAME == lowerCaseName) {
-            value.validateType(Top::TYPE);
+        } else if (TopK::NAME == lowerCaseName) {
+            value.validateType(TopK::TYPE);
             topK = value.getValue<int64_t>();
         } else {
             throw common::BinderException{"Unrecognized optional parameter: " + name};