--- a/extension/fts/src/function/fts_config.cpp
+++ b/extension/fts/src/function/fts_config.cpp
@@ -187,7 +187,7 @@ QueryFTSConfig::QueryFTSConfig(const function::optional_params_t& optionalParams
             isConjunctive = value.getValue<bool>();
         } else if (Top::NAME == lowerCaseName) {
             value.validateType(Top::TYPE);
-            top = value.getValue<int64_t>();
+            topK = value.getValue<int64_t>();
         } else {
             throw common::BinderException{"Unrecognized optional parameter: " + name};
         }