--- a/extension/fts/src/function/query_fts_bind_data.cpp
+++ b/extension/fts/src/function/query_fts_bind_data.cpp
@@ -48,7 +48,7 @@ QueryFTSConfig QueryFTSOptionalParams::getConfig() const {
             ExpressionUtil::evaluateLiteral<bool>(*conjunctive, LogicalType::BOOL());
     }
     if (topK != nullptr) {
-        config.topK = ExpressionUtil::evaluateLiteral<int64_t>(*topK, LogicalType::INT64());
+        config.topK = ExpressionUtil::evaluateLiteral<uint64_t>(*topK, LogicalType::UINT64());
     }
     return config;
 }
@@ -58,12 +58,12 @@ std::vector<std::string> QueryFTSBindData::getTerms(main::ClientContext& context
     FTSUtils::normalizeQuery(queryInStr);
     auto terms = StringUtils::split(queryInStr, " ");
     auto config = entry.getAuxInfo().cast<FTSIndexAuxInfo>().config;
-    auto stopWordsTable = context.getStorageManager()
-                              ->getTable(context.getCatalog()
-                                             ->getTableCatalogEntry(context.getTransaction(),
-                                                 config.stopWordsTableName)
-                                             ->getTableID())
-                              ->ptrCast<NodeTable>();
+    auto stopWordsTable =
+        context.getStorageManager()
+            ->getTable(context.getCatalog()
+                    ->getTableCatalogEntry(context.getTransaction(), config.stopWordsTableName)
+                    ->getTableID())
+            ->ptrCast<NodeTable>();
     return FTSUtils::stemTerms(terms, entry.getAuxInfo().cast<FTSIndexAuxInfo>().config,
         context.getMemoryManager(), stopWordsTable, context.getTransaction(),
         getConfig().isConjunctive);