--- a/extension/fts/src/include/function/fts_config.h
+++ b/extension/fts/src/include/function/fts_config.h
@@ -91,10 +91,10 @@ struct Conjunctive {
     static constexpr bool DEFAULT_VALUE = false;
 };
 
-struct Top {
+struct TopK {
     static constexpr const char* NAME = "top";
-    static constexpr common::LogicalTypeID TYPE = common::LogicalTypeID::INT64;
-    static constexpr uint64_t DEFAULT_VALUE = INT64_MAX;
+    static constexpr common::LogicalTypeID TYPE = common::LogicalTypeID::UINT64;
+    static constexpr uint64_t DEFAULT_VALUE = UINT64_MAX;
 };
 
 constexpr uint64_t INVALID_TOP_K = UINT64_MAX;
@@ -107,7 +107,7 @@ struct QueryFTSConfig {
     // document length.
     double b = B::DEFAULT_VALUE;
     bool isConjunctive = Conjunctive::DEFAULT_VALUE;
-    uint64_t topK = Top::DEFAULT_VALUE;
+    uint64_t topK = TopK::DEFAULT_VALUE;
 
     QueryFTSConfig() = default;
     explicit QueryFTSConfig(const function::optional_params_t& optionalParams);