--- a/extension/fts/src/include/function/fts_config.h
+++ b/extension/fts/src/include/function/fts_config.h
@@ -97,6 +97,8 @@ struct Top {
     static constexpr uint64_t DEFAULT_VALUE = INT64_MAX;
 };
 
+constexpr uint64_t INVALID_TOP_K = UINT64_MAX;
+
 struct QueryFTSConfig {
     // k: parameter controls the influence of term frequency saturation. It limits the effect of
     // additional occurrences of a term within a document.
@@ -105,7 +107,7 @@ struct QueryFTSConfig {
     // document length.
     double b = B::DEFAULT_VALUE;
     bool isConjunctive = Conjunctive::DEFAULT_VALUE;
-    uint64_t top = Top::DEFAULT_VALUE;
+    uint64_t topK = Top::DEFAULT_VALUE;
 
     QueryFTSConfig() = default;
     explicit QueryFTSConfig(const function::optional_params_t& optionalParams);