--- a/extension/fts/src/include/function/query_fts_bind_data.h
+++ b/extension/fts/src/include/function/query_fts_bind_data.h
@@ -12,6 +12,7 @@ struct QueryFTSOptionalParams {
     std::shared_ptr<binder::Expression> k;
     std::shared_ptr<binder::Expression> b;
     std::shared_ptr<binder::Expression> conjunctive;
+    std::shared_ptr<binder::Expression> top;
 
     explicit QueryFTSOptionalParams(const binder::expression_vector& optionalParams);