--- a/extension/fts/src/include/function/query_fts_bind_data.h
+++ b/extension/fts/src/include/function/query_fts_bind_data.h
@@ -17,6 +17,8 @@ struct QueryFTSOptionalParams {
     explicit QueryFTSOptionalParams(const binder::expression_vector& optionalParams);
 
     QueryFTSConfig getConfig() const;
+
+    bool isTopSet() const { return top != nullptr; }
 };
 
 struct QueryFTSBindData final : function::GDSBindData {