… src/binder/expression/expression_util.cpp