This PR adds the`top` optional parameter to `queryFTS` function. The `top` parameter chooses the `top k` result from the queryFTS output.

Performance improvement:
Query:
```
CALL query_fts_index('doc', 'contentIdx', 'dispossessed meaning', top:=topK) RETURN node,score order by score desc;
```

Result:
4 threads:

| topk  | Master | This branch |
|-------|------------------|-----------------|
| 50    | 2.808s           | 1.398s          |
| 500   | 2.977s           | 1.496s          |
| 5000  | 2.906s           | 1.728s          |